Move print_sep to utils
