+++ GPTifier/include/reporting.hpp
@@ -4,7 +4,6 @@
 
 namespace reporting {
 
-void print_sep();
 void print_error(const std::string &message);
 
 } // namespace reporting
